created python packages for arm_commander, and util in ob1_arm_control rospkg, refactored pytest test cases, added new fixtures and conftest.py and pytest.ini for test/fixture configurations. Reconstructed main.urdf with latest main.xacro
