+++ ob1_arm_control/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+pythonpath = scripts
