+++ ob1_arm_description/urdf/main.urdf
@@ -4,6 +4,13 @@
 <!-- |    EDITING THIS FILE BY HAND IS NOT RECOMMENDED                                 | -->
 <!-- =================================================================================== -->
 <robot name="robot">
+  <link name="ob1_arm_fixed_base_link">
+    <inertial>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+      <mass value="0"/>
+      <inertia ixx="0" ixy="0" ixz="0" iyy="0" iyz="0" izz="0"/>
+    </inertial>
+  </link>
   <link name="ob1_arm_base_link">
     <visual>
       <origin rpy="-1.5708 1.21431e-16 3.90606e-18" xyz="3.31402e-16 -0.0195 0.06"/>
@@ -471,6 +478,15 @@
     <joint_properties friction="0.0"/>
     <dynamics D="1" K="7000" damping="0.003" friction="0.0" mu_coulomb="0" mu_viscous="16"/>
   </joint>
+  <joint name="ob1_arm_fixed_base_joint" type="revolute">
+    <origin rpy="0 0 0" xyz="0 0 0"/>
+    <parent link="ob1_arm_fixed_base_link"/>
+    <child link="ob1_arm_base_link"/>
+    <axis xyz="0 0 1"/>
+    <limit effort="1" lower="-3.14" upper="3.14" velocity="20"/>
+    <joint_properties friction="0.0"/>
+    <dynamics D="1" K="7000" damping="0.003" friction="0.0" mu_coulomb="0" mu_viscous="16"/>
+  </joint>
   <link name="ob1_arm_gripper_base_link">
     <visual>
       <origin rpy="-1.31921 -1.52407 -0.251324" xyz="-0.04015 0.0144875 -0.0347028"/>
@@ -821,5 +837,11 @@
     <limit effort="1" lower="-3.14" upper="3.14" velocity="20"/>
     <joint_properties friction="0.0"/>
     <dynamics D="1" K="7000" damping="0.003" friction="0.0" mu_coulomb="0" mu_viscous="16"/>
+  </joint>
+  <!-- connect the arm to world with fixed joint -->
+  <link name="world"/>
+  <joint name="base" type="fixed">
+    <parent link="world"/>
+    <child link="ob1_arm_fixed_base_link"/>
   </joint>
 </robot>
